Consolidate shared test harness helpers and document flavors
Extract createD1FromSqlite and workers seed/schema factories into
test-support, adopt them in non-sibling-owned suites, and add the
node/workers/mcp-e2e/Playwright decision matrix to testing-principles.
