+++ packages/worker/src/entitlements/entitlements-service.workers.test.ts
@@ -7,33 +7,20 @@ import {
 	consoleWarn,
 	silenceExpectedConsoleWarns,
 } from '#worker/test-support/console-spies.ts'
+import { seedAccount } from '#worker/test-support/workers-seed.ts'
 import { createStableUserIdFromEmail } from '#worker/user-id.ts'
-
-async function seedAccount(
-	email: string,
-	plan: 'pro' | 'max',
-	stableUserId: string,
-) {
-	await env.APP_DB.prepare(
-		`INSERT INTO users (username, email, password_hash, email_verified_at, plan, stable_user_id)
-			VALUES (?, ?, ?, ?, ?, ?)`,
-	)
-		.bind(
-			`entitlement-service-${crypto.randomUUID().slice(0, 8)}`,
-			email,
-			'test-password-hash',
-			new Date().toISOString(),
-			plan,
-			stableUserId,
-		)
-		.run()
-}
 
 test('findUserAccountByStableUserId resolves accounts via indexed stable id and recovers from deletions', async () => {
 	await ensureEmailTestSchema(env.APP_DB)
 	const email = `reverse-lookup-${crypto.randomUUID()}@example.com`
 	const userId = await createStableUserIdFromEmail(email)
-	await seedAccount(email, 'pro', userId)
+	await seedAccount({
+		db: env.APP_DB,
+		email,
+		username: `entitlement-service-${crypto.randomUUID().slice(0, 8)}`,
+		plan: 'pro',
+		stableUserId: userId,
+	})
 
 	expect(await findUserAccountByStableUserId(env.APP_DB, userId)).toEqual({
 		email,
